Simplify API and framework linking.
Instead of NSURL, we now use an NSString to specify the path to the
script.

By default, set package.path in the Lua script (prior to running it) to
be the Resources directory in the main bundle.

The framework should be copied to the Frameworks directory in the
application bundle. The install path has thus been updated to
@loader_path/../Frameworks